+++ Source/Form/Gen4/Static4.cpp
@@ -174,11 +174,11 @@ void Static4::generatorCategoryIndexChanged(int index)
 {
     if (index >= 0)
     {
-        size_t size;
+        int size;
         const StaticTemplate4 *templates = Encounters4::getStaticEncounters(index, &size);
 
         ui->comboBoxGeneratorPokemon->clear();
-        for (size_t i = 0; i < size; i++)
+        for (int i = 0; i < size; i++)
         {
             if ((currentProfile->getVersion() & templates[i].getVersion()) != Game::None)
             {
@@ -295,11 +295,11 @@ void Static4::searcherCategoryIndexChanged(int index)
 {
     if (index >= 0)
     {
-        size_t size;
+        int size;
         const StaticTemplate4 *templates = Encounters4::getStaticEncounters(index, &size);
 
         ui->comboBoxSearcherPokemon->clear();
-        for (size_t i = 0; i < size; i++)
+        for (int i = 0; i < size; i++)
         {
             if ((currentProfile->getVersion() & templates[i].getVersion()) != Game::None)
             {
